+++ examples/examples_yaml_test.go
@@ -413,6 +413,17 @@ func TestYamlOidcIssuerExample(t *testing.T) {
 	})
 }
 
+func TestYamlPackageVersionExample(t *testing.T) {
+	cwd := getCwd(t)
+	digits := generateRandomFiveDigits()
+	integration.ProgramTest(t, &integration.ProgramTestOptions{
+		Dir: path.Join(cwd, ".", "yaml-package-version"),
+		Config: map[string]string{
+			"digits": digits,
+		},
+	})
+}
+
 func writePulumiYaml(t *testing.T, yamlContents interface{}) string {
 	tmpdir := t.TempDir()
 	b, err := yaml.Marshal(yamlContents)